Make test_neutron_dhcp_agent_list_hosting_net use net name from conf
The test_neutron_dhcp_agent_list_hosting_net test assumes the name of
the network hosted by the dhcp agent is 'private' which may not be the
case depending on the openstack installation. This patch changes the
test case to use the existing option 'fixed_network_name' which already
defaults to 'private'.

Fixes bug 1206694

Change-Id: I2c5ad2ff050f672a113643acbfe76c9360c3da8f