Add locks for all aggregates tests with hosts
This commit adds locking to all the methods in test_aggregates that
add a host to an aggregate. Nova change
I788782a9f21ec2672551f75123753175bb268586 added a check so that
a host being added to an aggregate is not already a member of an
availability zone. This can cause an issue with running tempest in
parallel because tests in test_aggregates uses only one host for
multiple tests. If any of these tests are run at the same time as a
test where the host is added to an availability zone it will fail.
By adding locking to all the tests with adding hosts to an aggregate
this ensures that this failure will not occur.

part of blueprint speed-up-tempest

Change-Id: Iaf8a903d92c88c7c3092eaa00a9f7a96e6348ddc