Skip secgroup rules invalid id tests until bug 1182384 is fixed
There are some existing reviews which try to deal with bug 1182384:

nova: https://review.openstack.org/#/c/35285/

tempest: https://review.openstack.org/#/c/29899/

There is also a mailing list discussion about this (see the link in the
bug report).

Until there is consensus on how to fix this, skip the compute tests if
neutron is available.

Change-Id: I3f0d21c23661f556354c321476ba559d4925dccd