Skip more security group tests until bug 1182384 is fixed
This patch builds on change I3f0d21c23661f556354c321476ba559d4925dccd
which skipped a couple of tests that are failing in nova if neutron is
configured. The previous patch missed a couple of other tests that fail
for the same bug. This patch skips those other failing tests for the
same bug.

Related-Bug: #1182384

Change-Id: I4f191f21232f9e58ad09456261fb6247d64f22b5