Rework class inheritance for scenario tests
This commit reworks the class inheritance used by the scenario tests
to make is a bit simpler and similar to what is used for the api
tests. The scenario tests now inheret from the base test case directly
and client creation is handled directly instead of looping over the
client_attr_names list.

Change-Id: Id8bf2db3ec8f706ca466b86fa71df3ee2607fe88