Log request ids from glance and nova
This commit adds an option a log message at the info level to
print the req-id returned from nova and glance. This can be useful
when trying to debug test failures because it provides you with a
unique identifier in the nova and glance logs for request
information. This is done at the info level so that it will be
included in the subunit stream to be printed out during a test
failure to help with debugging.

Fixes bug 1210639

Change-Id: I819f94c538e6a15732dbb4b5c2dafdc270e0b2b4