Cleanup: Add common "create_server" method
In scenario tests, there are many "create server" methods and their
behaviors are almost same. They creates a server with random name,
default image and default flavor, and waits for server status "ACTIVE".
So this patch adds a common "create_server" method and uses the method
instead of each implementation for code-cleanup.

Change-Id: I663cf1c994307160c9f0ebf51fa4c920f2b6125e