Add nose to run_tests and tox for python 2.6
This commit adds python 2.6 to both tox and run_tests. 2 new tox jobs
are added py26-full and py26-smoke, which will use nose as the test
runner. run_tests.sh is also modified to detect whether python2.6 is
being used, and if so it will use nose instead of testr.

Change-Id: Ia7cecbea99ba2f995a726def6f775a313c4e94ba