Fix skip tracker regex for multi-line skips
This commit changes the regex used for detecting a skip line by
removing the requirement on an '(' immediately following 'skip'.
This changes the skip tracker to match multiline skips (assuming
'skip' is on the same line as the bug number. It also enables
conditional skips that have a bug number from being picked up.

Change-Id: I95a0521661084034f7625c3c517ae9210120de4c