Cleanup: Add common "create volume" method
In scenario tests, there are some "create volume" methods and their
behaviors are almost the same. They create a volume with random name,
size 1GB and wait for volume status "available".
So this patch adds a common "create_volume" method and uses the method
instead of each implementation for code-cleanup.

Change-Id: I06f29a77116fd617f47b0ae6916de933ebb7fdc3