Initial basic setup of openstack and tempest config file
The objective of this script is to configure openstack environment +
tempest config file while relying on nothing but the current Openstack
installation and/or existing tempest config file.
it achieves this by using only Openstack python clients and/or global Openstack
environment variables and for some tempest specific settings, the tempest
config file.
This script is a WIP and currently support only creating users and configure
the tempest config file with active images found in openstack.

TODO:
    1. Add support for existing config and not rely on sample only
    2. Add download & creation of images
    3. Add network configuration
    4. more advance features as required

Change-Id: I06f38280e6b550d03c64954e042aee7331ecb31b