Update requirements from global requirements
This commit just syncs the tempest requirement list with what's in
the global requirements repo. We are already using them for running
in the gate and with devstack so there isn't a reason to update the
lists in the tempest repo. This commit also adds jsonschema to the
list of requirements. It was removed at one point because it caused
an issue with grenade but since the global requirements fix ups this
is no longer an issue.

Fixes bug 1205980

Change-Id: I406eddc0b2363c98035edef6fc651228a2bdd04f