Fix ssh timeout issue
SSH client is using a socket timeout and an execution timeout. For
command execution, if the executed command did not generate output
within the channel timeout, test dies with TimeoutException. This is
misleading, because it wasn't the command that timed out, it just did
not generate output within channel timeout.

fixes bug 1215794

Change-Id: I90b9cb7c9cebcb78655b6c1a46f9fb09ff7deb5e