Handling network resources in tenant isolation
Functionality is added to create and delete isolated network, subnet and router
when isolated credentials are created and deleted for a primary, admin or alt
tenant.

When a set of isolated credentials are created, the tenant gets also a private
network, subnet and router. The router connects the tenat's network to the
public network designated by public_network_id in etc/tempest.conf

Change-Id: I7587c85017cca09f7a67eae0670f67b2bceacb60
Fixes: bug #1216076