Fix test_admin_catalog_list
This commit fixes test_admin_catalog_list() so that it tests more
specific details of the admin catalog. It also removes the test's
dependency on each table in the output having a service label. This
dependency was occasionally causing a failure if the test ran at the
same time as the keystone api tests.

Fixes bug 1217159

Change-Id: I4446abb86a3f1f6514fc9de1d00c06f82943b9d9