Add handling for inherited stress attributes
In some cases inheritance is used for test cases (like XML/JSON tests).
For many of those cases inherited stress attribute are misleading
since they will execute nearly the same load on the system.

To avoid this issue we skip all inherited stress attributes by default
and add parameters to deactivate this on a global level and function
level.

Change-Id: I4260844ed15917106436d836c939157e32a59554
Fixes: bug #1219402