Add device name, ssh password to Tempest config
The following options have been added to the compute section of the
Tempest configuration:

- image_ssh_password
- image_alt_ssh_password
- volume_device_name

The "volume_device_name" is being added to rid the test_attach_volume
test of hardcoded data. In the effort to remove a hardcoded password
from the same test, we introduce the option "image_ssh_password" to
store the password instead. Then, the deprecated "ssh_user" in the
test is swapped out in favor of "image_ssh_user". Finally, we add an
"image_alt_ssh_password" so the option remains symmetrical to the
"image_ssh_password" option.

Closes-Bug: #1220514
Change-Id: If7816e534200c54826c1da0d0464f643163b8657