Update docs config
This commit updates the docs conf.py to make it more closely resemble
what is used it other projects(based on nova's equivilent file). It
also removes listing Sean Dague as the sole author for the Tempest
docs because it is a collaborative effort. It also fixes a pep8
failure on the imports.

Change-Id: I535858dc9d413e8d5f92d44194137f615390d117