Move _check_tenant_network_connectivity to the end
For enabling tenant network connectivity checks when tenant networks
are not reachable, connectivity will be checked verifying if an
instance can be reached from another instance on the same network.

In order to do so, the instance from which the command is issued
must be accessible from the tempest test runner; and this will be
achieved associating a floating IP to it, which is exactly what
_check_public_network_connectivity does. By executing it first,
the floating IP association will be already available when
checking tenant network connectivity.

Partial-Bug #1226158

Change-Id: If2eea34e363ed40e9a0b38173534184d02a9078f