Inject "-tempest-" string to rand_name
In many tempest tests, we use rand_name(). However, these generated
names are very general. So it is difficult to distinguish the resource
is generated from Tempest or not.
And, in many tests, we use rand_name() with '-' suffix such as
rand_name('foo-'). I think this is not smart.
So this commit injects "-tempest-" to rand_name().

And, by this commit, the generated rand name will include "--" such as
'foo--tempest-123'.
So I will fix them in another commit.

Change-Id: I7bb8f5bbe2346b0b0ab6e91757e20a67b9098f8a