Add a function for creating a server on a given network
For enabling tenant network connectivity checks when tenant networks
are not reachable, connectivity will be checked verifying if an
instance can be reached from another instance on the same network.

To this aim it will be necessary to boot additional servers.
This patch refactors _create_servers extracting the code for starting
a single server into its own method.

This method will allow the caller to choose on which networks the
server should be booted. This will be useful for other useful test
cases such as verifying tenant connectivity across distinct networks
connected by a router.

Partial-Bug #1226158

Change-Id: I0051eb1a2ef4113a624517b15775c9c0d7bfef1d