API tests for neutron router gateway
This patch adds tests for validating the Neutron API for managing
router gateways.

To run these tests successfully, the neutron client has been augmented
with a method to send the enable_snat attribute to the server; this
method should be invoked only when the client has admin credentials;

Also, this patch adds the capability of specifying filter when
querying the API for ports, as this is needed by gateway tests.

Change-Id: I4e92384559bb2093c0eddf78f5dbb526af1acc38
Closes-Bug: #1228091