Add locking to test_list_hosts_with_zone
test_list_hosts_with_zone() makes 2 list_hosts calls and expects that
the hosts' availability zone will remain constant between the calls.
However when running in parallel there is no guarantee of this, and
the aggregates tests will change the compute host's availability
zone. This commit adds the 'availability_zone' to the test to ensure
it doesn't run at the same time as the tests which will change the
host's availability zone.

Fixes bug 1228225

Change-Id: I14773e38b5cc35679b16940ad230c7c81aeaad9c