Sync install_venv_common.py from oslo-incubator
This is needed to ensure that when using run_tests.sh requirements
are handled correctly. Both requirements.txt and test-requirements.txt
need to be handed to pip at the same time to ensure that all
dependencies are satisfied and we don't upgrade a package to be
outside of the version constraints set by requirements.txt when
installing dependencies for the packages in test-requirements.txt.

Change-Id: I89bcf4e44b7274dca535e8e1d2eeb0adb142cabb