Fix json version of bootable check
This is a follow up to change id:
  I8a4b1f4ce3a6a420e8152ea8c66a9b3fc127fe96

The initial change only addressed the xml aspect of the bug
and ignored the json tests.

This change updates the check to catch the same errors when
performing the json version of the checks.

Both of these patches are temporary and after the Cinder update
lands should be replaced with a proper/strictly enforced check of
'true' and 'false' values in the volume['bootable'] field.

Change-Id: Iffe5d380410f5c698018a5c3106c99af05f0b80d