Explicity specify network for heat slow tests
The heat slow test NeutronResourcesTestJSON is causing problems
for the other heat tests when run concurrenty.

Because NeutronResourcesTestJSON creates a network and the other tests
do not specify any, an amibiguity is created. This means that nova
server creation fails due to not knowing which network to attach to.

This change explicity launches servers with the network specified in
config compute.fixed_network_name.

This may mean that these tests cannot be run with nova networking.
This is most likely not a problem as there are currently no plans
to gate heat-slow tests on nova networking.

Change-Id: I16d50271a36c9bc9d0c4fb0588ae75ba100b145f