Handle test_list_servers_by_changes_since without tenant isolation
The test_list_servers_by_changes_since test fails when running with
allow_tenant_isolation=false because there is an arbitrary 2 minutes to
determine the timedelta for the changes-since in the query. That doesn't
work when the servers created in the test are all under the same
user/tenant because the tearDown doesn't block until the servers are
deleted.

This patch makes the timedelta code smarter in that it narrows the
changes-since down to when the test case started setting up the servers
for the test and when it does the query.

Closes-Bug: #1195309

Change-Id: Ie848637ea490da2151e5db8292726111d45a8c2d