Cleanup test_list_server_filters setup/teardown
The test_list_server_filters test class was doing a lot of
setup/teardown work that is now handled in the parent class. We still
have to explicitly set the name since some tests use it for filtering
but otherwise we can rely on the parent class to do most of the heavy
lifting here.

This change makes it so that the instances have a name based on the test
class rather than just a random integer which is hard to use for
debugging leaked instances.

This change also removes the get call from setup and moves it into the
test cases that actually need to use it (which are currently being
skipped anyway).

Closes-Bug: 1235411

Change-Id: I66813c900059dd4232de645c225032e58e9aecde