Use predictable instance/volume names in test_volumes_actions
The setup method of test_volumes_actions generates a random name for the
instance and volume it uses for the tests but the names don't include
the test class so if they don't get cleaned up it's hard to debug where
they came from.  This patch simply adds the class name to the
instance/volume name generation.

Closes-Bug: #1235411

Change-Id: If457e73c11f08fd57ee77a85208de84fc39eda8c