Script to filter logs for ERRORs based on whitelist
The script will take a directory or url containing log files.
For now all non-whitelisted errors will be dumped to the console but
the script will always return success. Once we are convinced it is reliable
enough we can change it to fail on non-whitelisted errors.

Partially implements blueprint fail-gate-on-log-errors

Change-Id: I30b0eee1055f47aaad7984d886c739ccf5aa6186