Fixing ImagesOneServerTestXml issue
Now operations started with the server when it's task state was not
None. Ie. Idle / 'ready for action'.

The wait for wait_for_server_status method implicitly waits for the None
task state, it should be able to fix the issue.

The server will be rebuilded if otherwise it cannot be move to a clean active
state, this issue will be logged.
It is an unexpected exception, so the log verifier should be able to
report it.

Fixing bug 1218391

Change-Id: Ib0e90a06e0e34f134bcca74712939e87e235c1ea