Update to latest pbr
This commit moves switches tempest from the d2to1 version of pbr and
move to the latest version of pbr. However, there is a bug with
build_spinx so it doesn't work with a doc version set so that section
is removed from the doc conf.py.

Change-Id: I4fe14eb2b2e81132f8fa99936a031176b38c437b
(cherry picked from commit 271139aa7f1fa5f8d9e663702bf275378456fa46)