Add config feature verification script
This commit adds a new tool to tempest that will verify that API
queryable config options are set correctly. Right now the list of
options that are verified is very short. Later on additional checks
will be added to verify other services features when the tempest
clients for the other services get functions to query enabled
extensions.

Partially Implements: blueprint config-verification

Change-Id: Ie3b5fadd74460fb7199bff7046ec0efe37268b94