Fix precondition checks for test_network_basic_ops
Running tempest with quantum disabled would previously raise
skip exceptions indicating that one of public_router_id or
public_network_id was required.  These values are not required
if quantum is disabled.  This change ensures that the presence
of these values will only be checked if quantum is enabled.

The master branch already behaves this way, but the behavior
there relies on a refactor that would have been difficult to
backport.

Change-Id: I4eee9403b219fd97d16faae0ed2b11348edbc04e
Closes-Bug: #1243593