Add sample config check to tox pep8 job
This commit takes the check_uptodate.sh script from nova and adds it
to the pep8 tox job. This will enforce that the sample config file is
up to date for each proposed commit. This also ensures that config
options are setup properly and that the style in the sample config is
consistent because it's auto-generated.

Partially Implements: blueprint config-cleanup

Change-Id: I9efaa9fc1cf09e95d7111b000895c7b49705713b