Add special serial smoke option
We want smoketests to be parallel by default (for grenade etc), but that won't work
for neutron, so add a special case. Since we don't want to break
anything making this switch, it has to be done in 3 steps:

1) this patch
2) make neutron use smoke-serial
3) change smoke to be parallel

Note this is a backport to grizzly so the parallel options are not being
added, we are just maintaining serialized testing on the grizzly branch.

Change-Id: I03ded120c0b496026e39f14f6715624a91ac7558