Skip test_create_backup
test_get_console_output fails many times now.
test_create_backup is a previous test of test_get_console_output, and
test_create_backup has been merged since 2013/11/15. The date is when
this failure started to happen.

But this is just a inference. However, after much checking the problem
seems to go away with this change.

Change-Id: I7fbdcec6c7990d7a9822037782729814cc6b6b79