Forced isolation for the tests in test_list_servers_negative.py
In bug 1252971, race condition will arise if the allow_tenant_isolation
is false. ListServersNegativeTestJSON and ListServersNegativeTestXML
will share the tenant, if both of them check if there are existing servers
or not before both of them create new servers, the actual servers will
be double of the expected.
this also removes the redundant code in this files.

Closes-bug: 1252971

Change-Id: I2f859057a71ccada86b5733c1a27386fcd6bc4b5