Use channel_timeout for SSH connection timeout
Occasionally, SSH will get wedged so that a connection attempt is stuck
forever. When this happens, we need Tempest to abort the attempt and try
again. Currently, the individual connection timeout is set to the
overall timeout, so there will only ever be one attempt if this happens.
Using the channel_timeout instead will ensure that multiple connection
attempts are made even when the connection is wedged.

Fixes bug 1236524

Change-Id: Ie8dff41780bbf004cff5c880db202a8ae23a85c1