Revert "Add test_cases for cinder cli in v2 version"
These tests aren't parallel safe, as they make calls into cinder
that assume nothing else is changing the environment, but they
are. This was seen in errors like:

http://logs.openstack.org/43/59243/2/gate/gate-tempest-devstack-vm-postgres-full/e2be8a0/

Doing a full revert as this is currently bouncing a lot of code
that is unrelated. This will need to be rethought in a future
iteration.

This reverts commit cceda7d4c727815aff95d8deb941ffce069f2c8b.

Change-Id: I44ef405695074615ef6447078958791af9b99827