Add a control point for floating IP assignment
In the basic network scenario test, the test for accessiblity via the
floating IP stands a very good chance of failing because the attempt to
connect begins before the system has had a chance to associate the
floating IP address.  While an extended delay does signify a performance
issue for those types of operations, the timing of the test can be
adversly affected by other factors (e.g. parallel tests, loaded test
  system). If it is extreme the check for the IP address will fail the
test as well. This addition also covers a loophole in the test where
the connectivity would work but the metadata that is provided through
the client call is not updated properly.

An additional small fix in manager.py resolves an issue when a client
API uses mixed or lower case for status strings (e.g. cinder).

There isn't a specific bug that this modification addresses.

Change-Id: Ifb1889f67588e6ee4d19f592da61ab574d456be2