Add unit tests for rest_client
This commit adds the first unit tests for the rest client. This is just
the first basic set of tests which were written to mostly verify some
common fake objects that will be useful moving forward. The intent is
for more advanced to be built off of this.

Partially implements bp unit-tests

Change-Id: I19173523ceec3c4fce62f9a3ab61cc9ee23c1801