Separate negative tests for test_availability_zone
Move negative tests from test_availability_zone.py to
test_availability_zone_negative.py

This patch also uses 'AZ' short for the 'AvailabilityZone'.
This is because if the class name 'AvailabilityZoneAdminV3
NegativeTestJSON' is used, the tempest test will occur an
error 'Project name should not be greater than 64 characters.'
The Project name is made of class name and the characters
'-tempest-xxxxxxxxxx-tenant'.
'-tempest-xxxxxxxxxx-tenant' is 26 characters, so the class
name should not exceed 38 characters.

Partially implements blueprint negative-test-files
Change-Id: Ic72264e89ac5fd700202b0c75cc63116d8fca172