Remove generic_setup_package() function
The generic_setup_package() function was previously run at init which
was causing a dependency on a config when a file was imported. This
commit takes that functionality and puts it in a classmethod and calls
that in setUpClass. There was nothing in generic_setup_package() that
needed to be run during so init, so this moves it to the proper place.

Change-Id: I32fa0fb1e230a98a3a9ca732df2d013ca4f13dce