Start instances using fixed network when possible
Uses the fixed_network_name option to select which network
to boot test instances on.  This allows Tempest to be run
in environments where there are multiple networks available.

This problem does not apply to nova-network environments because
it will happily start instances with multiple networks attached,
so the behavior is left as before in that case.

Co-Authored-By: Ben Nemec <bnemec@redhat.com>
Co-Authored-By: Vincent Untz <vuntz@suse.com>
Co-Authored-By: git-harry <git-harry@live.co.uk>

Change-Id: Ib5b84b59e3d182d8b9cc83954537c32f3eb9e388
Closes-Bug: #1250866