Wait for backup images to be ACTIVE in test_create_backup
The test_create_backup test creates two backup snapshot images of the
instance right away and waits for the instance to be ACTIVE before
continuing each time, and before creating the third and final backup
image.

The problem is the instance is always ACTIVE, it's the task state that
is transitioning on the instance while it's creating backup snapshot
instances and checking for old rotations to delete.

The test needs to not only check that the instance is ACTIVE before
creating each new backup but also that the snapshot image backups
themselves are ACTIVE which is a better indication that the instance's
state transition is complete.

Closes-Bug: #1257561

Change-Id: I5b423b3424e1dfe6cf5f8fef0dc9538961c1f28d