allow hypervisors to be down but still pass
This test case is pretty raw, but it does make some invalid
assumptions about using the first hypervisor for everything. In
the uptime case it really needs to loop over all of them otherwise
if the first hypervisor is disabled this will fail.

This also does a cleanup refactoring of the hypervisor list check.

Change-Id: I9c3db794269f61aaaa55d432b14edf2d0fd1c774
Closes-Bug: #1193194