Catch ssh exception and detach in min scenario
The test_minimum_basic test goes through a number of tests however
doesn't do any clean up if a failur occurs.  The master clean-up
routine is ok, but it doesn't account for things like attached volumes.

One issue here is that the ssh_to_server times out do to some other
problem and raises an SSHTimeout exception.  This exception isn't dealt
with, the test dumps out and tries to cleanup by deleting the instance
and the volume, but the volume-delete will fail because it's status is
still in-use.

This exposes another problem, it seems that nova delete will clean up
an instance but fairly regular neglect to detach the volume. This patch
add an addCleanup() call to nova.volume_detach.

Change-Id: I76d19f096d6cb4122eea2e7e5de4eb37b97a4b87
Partial-Bug: #1213215