Add test_discover module to provide a load_tests hook
This commit adds a load_tests hook that will be used to discover the
set of tempest tests that excludes the unit tests. We can't just
discover from the tempest topdir like we previously did because of
mutually exclusive requirements between the normal tempest tests and
the unit tests. (for example mox and mock should not be required for
a tempest run)

Change-Id: I562f033d78e654675fd77af3bf64c44945127bd7