Delete a BuildError server
If a server creation fails with create_test_server() in setUpClass,
the server would not be removed because tearDownclass is not called.
As the result, many BuildError servers are created and "Quota exceeded"
happens if server creations continue to fail.

This patch makes create_test_server() removes a server when BuildError
happens for avoiding "Quota exceeded" exception and investigating the
original problem which is due to the BuildError.

Change-Id: I55bb386bbea550de11cc7a81c5d056680057523b
Related-Bug: #1257641