remove last errant file parse warning
provide a way to be able to instantiate the config object without
parsing a config file. This is expected to only be used for the
config file generator comparitor.

This means we finally don't generate warning messages on config
file parsing during our pep8 runs, and we do the right thing and
explode during tests if the config file doesn't exist in a real
way.

Change-Id: Ic4ae22814e2fdafc89a1997471283ccf8f3663bc