Use wrapper create_volume() in volume tests
There is wrapper create_volume() in base.py of volume tests, the method
creates a volume with 1GB size in default and registered the volume for
deleting it when tests finish.
However, the method is not used at many places and similar codes exist
for the same purpose.

This patch makes the volume tests use this wrapper create_volume() for
cleaning up. Some tests do not use wrapper create_volume() after applying
this patch, because these tests are for testing create_volume API and
it is better to call the API directly without wrapper method.

Change-Id: I6d18d92984119f137801b56f3442d0c5851e0ef1