Remove Swift container-sync test skipping
The skipping reason was bug #1093743 that
is currently marked as invalid in Swift bugtracker.

This patch remove the skip and propose some
refactoring to the test code.

The test is marked as slow as it can take long
to run as it depends on container-sync process run
internal which is 300 seconds by default.

Change-Id: I939a014646b42cb2a461f874a5ad0edfc79c5561