Add neutron extension support to verify_tempest_config
This commit adds support for the neutron extension detection to the
verify_tempest_config script. The neutron extensions api behaves
slightly differently from the other services in that several of the
extensions' name field is really a title that includes spaces. This
can't be used for the list option and the tempest code already
assumes that the alias field is used instead of name.

Partially implements bp config-verification

Change-Id: Ib60f191fec1cb6b9d60055a848749b07b8393cf5