Fix logic for config file env variables
This commit fixes the logic around the 2 env variables to set the
location of the tempest config file. Previously the check to see
whether the config variables should be used instead of the default
path would be too selective and the default path etc/tempest.conf
would be used. The only check needed is whether the specified file
exists or not. The presence of the env variables shouldn't affect
that logic.

Change-Id: Id6be0f067673d600bf04dea700e69c2baffb9d9f