Fix class inheritance for volume_types tests
Previously the tests in test_volume_types.py manually setup an
admin_client using the admin credentials specified in the config file.
However, this is not correct because with tenant isolation those
credentials shouldn't be used other then to create a new tenant.
Additionally, everything performed in the test class's setUpClass was
already done in the base class. This commit corrects this by
changing the test class inheritance to the proper base class.

Closes-Bug: #1271312

Change-Id: Ia798c2c41d132a2ed17d6e0a9215d21a00c97a32